Refactoring in progress... (seems to work, but UpdateNewItem still buggy)
